chore: organize source files under "/src" dir
"/src" is a traditional directory under which to store source
files. Move all source files except the driver file to "/src"
and update all import statements.
