+++ src/__init__.py
@@ -1,0 +1,11 @@
+import src.AI_Taxonomy
+import src.csv_pull
+import src.csv_push
+import src.database_init
+import src.DatabaseManager
+import src.generateAST
+import src.github_pull
+import src.identifiers
+import src.store_result
+import src.symbolTable
+import src.tokens
